moves feature information into its own component
