+++ src/groups/RouteNetworkDiagram/FeatureDetailContainer/FeatureDetailContainer.scss
@@ -11,8 +11,8 @@
     border-radius: 4px;
 
     .feature-details-info {
-      font-size: 16px;
-      padding: 10px;
+      font-size: 18px;
+      padding: 8px;
       flex: 1;
       line-height: 1.5;
