+++ src/groups/RouteNetworkDiagram/FeatureInformation/FeatureInformation.scss
@@ -1,0 +1,14 @@
+.feature-information {
+  display: flex;
+  padding: 20px 10px;
+  justify-content: center;
+
+  .feature-informations {
+    display: flex;
+    flex-wrap: wrap;
+    max-width: 1000px;
+    justify-content: space-between;
+    width: 100%;
+    font-size: 18px;
+  }
+}
