+++ src/groups/RouteNetworkDiagram/FeatureInformation/FeatureInformationGql.ts
@@ -1,0 +1,29 @@
+export interface QueryRouteNetworkElementResponse {
+  routeNetwork: {
+    routeElement: {
+      routeNodeInfo?: {
+        function?: string;
+        kind?: string;
+      };
+      namingInfo?: {
+        name?: string;
+      };
+    };
+  };
+}
+
+export const QUERY_ROUTE_NETWORK_ELEMENT = `
+query ($routeElementId: ID!) {
+  routeNetwork {
+    routeElement(id: $routeElementId) {
+      routeNodeInfo {
+        function
+        kind
+      }
+      namingInfo {
+        name
+      }
+    }
+  }
+}
+`;
