+++ src/groups/RouteNetworkDiagram/FeatureInformation/index.ts
@@ -1,0 +1,4 @@
+import FeatureInformation from "./FeatureInformation";
+import "./FeatureInformation.scss";
+
+export default FeatureInformation;
