+++ src/groups/RouteNetworkDiagram/RouteNetworkDiagram.scss
@@ -1,18 +1,3 @@
 .route-network-diagram {
   width: 100%;
-
-  .feature-information-container {
-    display: flex;
-    padding: 20px 10px;
-    justify-content: center;
-
-    .feature-informations {
-      display: flex;
-      flex-wrap: wrap;
-      max-width: 1000px;
-      justify-content: space-between;
-      width: 100%;
-      font-size: 18px;
-    }
-  }
 }
