+++ src/groups/RouteNetworkDiagram/RouteNetworkDiagramGql.ts
@@ -249,36 +249,6 @@ mutation (
 }
 `;
 
-export interface QueryRouteNetworkElementResponse {
-  routeNetwork: {
-    routeElement: {
-      routeNodeInfo?: {
-        function?: string;
-        kind?: string;
-      };
-      namingInfo?: {
-        name?: string;
-      };
-    };
-  };
-}
-
-export const QUERY_ROUTE_NETWORK_ELEMENT = `
-query ($routeElementId: ID!) {
-  routeNetwork {
-    routeElement(id: $routeElementId) {
-      routeNodeInfo {
-        function
-        kind
-      }
-      namingInfo {
-        name
-      }
-    }
-  }
-}
-`;
-
 export interface RemoveSpanStructureResponse {
   spanEquipment: {
     removeSpanStructure: {
